backend/requirements.txt
crewai==0.30.11
flask==3.0.0
flask-cors==4.0.0
flask-jwt-extended==4.5.3
flask-bcrypt==1.0.1
pymongo==4.6.1
python-dotenv==1.0.0
pypdf==3.17.4
pdfplumber==0.10.3
spacy==3.7.2
python-dateutil==2.8.2
email-validator==2.1.0
werkzeug==3.0.1
bson==0.5.10
langchain==0.1.0
langchain-openai==0.0.5
openai==1.6.1 